requirements.txt
azure-functions
azure-storage-blob
pandas
pyarrow
azure-identity
openpyxl